requirements.txt
django==4.0.3
joblib==1.1.0
numpy==1.22.2
pandas==1.4.1
python-decouple==3.6
scikit_learn==1.0.2
spacy==3.2.3
psycopg2
Pillow